requirements.txt
pyspark
mysql-connector-python
jdk 11 :https://drive.google.com/drive/folders/1YiKNzQhiNOz_S_gCyjd_IbWwcAAnvqlx?usp=sharing

